Move LocalizedStrings.Console and .Glowstone to outer level
Makes references to localized strings more concise, and reduces the depth of nesting.
